TimerPage created to get reading time and save Reading events by book

+++ lib/dashboard/domain/models/events.dart
@@ -6,7 +6,7 @@ class LibraryStatisticsConsulted extends LibraryEvent {
   List<Book> inProgressBooks;
   List<Book> pendingBooks;
   List<Book> finalizeBooks;
-  int readingHours;
+  double readingHours;
   double pagesMerMinute;
 
   LibraryStatisticsConsulted({required this.pendingBooks, required this.finalizeBooks, required this.inProgressBooks, required this.pagesMerMinute, required this.readingHours});
